Add Vim, Yarn and Update Aliases
This commit adds basic inital packages for Vim -> the vim bindings to
allow for this commit message to automatically wrap.

It also adds pathing for Yarn. Yarn will be added to the Brewfile later
on, once it is determined to best way to actively keep it updated.

This also adds custom coloring for ls commands on the machine. It will
now first look for an Iterm2 theme if possible, if not then resort to
standard colors for ls.