Add gcloud autocompletion
This commit will add autocompletion scripts to gcloud in my dotfiles.
The scripts are pulled and attributed from authors on Github. I decided
not to include an installer for gcloud, as it seems too much work for
the benefit of only needing it on my work-development machine. This is a
good example for needing multiple branches/a way to manage the dotfiles
slighty more efficiently.